Split runner shell dispatch tests
